Move SuspensionFunction to annotations package
